// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  CUSTOMER: 'CUSTOMER',
  COURIER: 'COURIER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const MenuCategory = {
  NASI: 'NASI',
  LAUK: 'LAUK',
  BAKSO: 'BAKSO',
  SOTO: 'SOTO',
  AYAM: 'AYAM',
  MIE: 'MIE',
  MINUMAN: 'MINUMAN',
  SNACK: 'SNACK',
  OTHER: 'OTHER'
} as const

export type MenuCategory = (typeof MenuCategory)[keyof typeof MenuCategory]


export const OrderStatus = {
  ORDERED: 'ORDERED',
  PAYMENT_PENDING: 'PAYMENT_PENDING',
  PROCESSED: 'PROCESSED',
  ON_DELIVERY: 'ON_DELIVERY',
  READY: 'READY',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const DeliveryMethod = {
  DIANTAR: 'DIANTAR',
  AMBIL_SENDIRI: 'AMBIL_SENDIRI'
} as const

export type DeliveryMethod = (typeof DeliveryMethod)[keyof typeof DeliveryMethod]


export const PaymentStatus = {
  PENDING: 'PENDING',
  VERIFIED: 'VERIFIED',
  PAID: 'PAID',
  HANDED_OVER: 'HANDED_OVER',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const PaymentMethod = {
  CASH: 'CASH',
  QRIS: 'QRIS',
  GOPAY: 'GOPAY',
  SHOPEEPAY: 'SHOPEEPAY',
  OVO: 'OVO',
  DANA: 'DANA',
  TRANSFER: 'TRANSFER'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const PermissionAction = {
  CREATE: 'CREATE',
  READ: 'READ',
  UPDATE: 'UPDATE',
  DELETE: 'DELETE',
  MANAGE: 'MANAGE'
} as const

export type PermissionAction = (typeof PermissionAction)[keyof typeof PermissionAction]


export const PermissionResource = {
  MENU: 'MENU',
  ORDER: 'ORDER',
  CUSTOMER: 'CUSTOMER',
  COURIER: 'COURIER',
  PAYMENT: 'PAYMENT',
  ANALYTICS: 'ANALYTICS',
  SETTINGS: 'SETTINGS',
  PROMO: 'PROMO',
  REVIEW: 'REVIEW',
  NOTIFICATION: 'NOTIFICATION',
  PRIVILEGE: 'PRIVILEGE'
} as const

export type PermissionResource = (typeof PermissionResource)[keyof typeof PermissionResource]


export const NotificationType = {
  NEW_ORDER: 'NEW_ORDER',
  ORDER_STATUS: 'ORDER_STATUS',
  PAYMENT_SUCCESS: 'PAYMENT_SUCCESS',
  PAYMENT_FAILED: 'PAYMENT_FAILED',
  PROMO: 'PROMO',
  SYSTEM: 'SYSTEM'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]


export const PromoType = {
  PERCENTAGE: 'PERCENTAGE',
  FIXED_AMOUNT: 'FIXED_AMOUNT',
  FREE_DELIVERY: 'FREE_DELIVERY'
} as const

export type PromoType = (typeof PromoType)[keyof typeof PromoType]